Move enums into schema
